+++ src/primitives/index.ts
@@ -1,6 +1,7 @@
 export * from "./alert";
 export * from "./avatar";
 export * from "./badge";
+export * from "./brand-icon";
 export * from "./button";
 export * from "./card";
 export * from "./checkbox";